[refactor] migrate ThermoMicroscope to standalone autoscript.py file
