+++ fibsem/utils.py
@@ -282,7 +282,6 @@ def setup_session(
         configure_logging(session_path, _DEBUG=debug)
 
     # connect to microscope
-    import fibsem.microscope as fibsem_microscope
 
     # cheap overloading
     if ip_address:
@@ -295,7 +294,8 @@ def setup_session(
     ip_address = settings.system.info.ip_address
 
     if manufacturer in VALID_THERMO_FISHER:
-        microscope = fibsem_microscope.ThermoMicroscope(settings.system)
+        from fibsem.microscopes.autoscript import ThermoMicroscope
+        microscope = ThermoMicroscope(settings.system)
         microscope.connect_to_microscope(
             ip_address=ip_address, port=7520
         )
